Rename git completion file to not conflict with fish
fish ships with a default git.fish file with git completions.
When this is installed, it overrides that, and since this relies
on functions defined in that default file itself, e.g. __fish_git_branches,
it doesn't work.

By renaming it to git-flow.fish we ensure it is scoped correctly
and works as expected.